umi percentile

+++ test_cases/fix_stats.py
@@ -1,9 +1,12 @@
 from pathlib import Path
 import shutil
 
-for p in Path('test_cases').glob('**/counts.tsv.stats.tsv'):
-    if 'actual' in str(p):
-        print(p)
-        target = p.parent.parent.parent / "output/counts.tsv.stats.tsv"
-        print(target)
-        shutil.copy(p, target)
+
+for fn in 'counts.tsv.stats.tsv', 'matrix.mtx.stats.tsv':
+
+    for p in Path('test_cases').glob(f'**/{fn}'):
+        if 'actual' in str(p):
+            print(p)
+            target = p.parent.parent.parent / f"{p.parent.parent.name}/{fn}"
+            if target.exists():
+                shutil.copy(p, target)
